requirements.txt
# Web 框架与服务器
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic-settings>=2.0.0
pydantic>=2.0.0

# LLM 与工作流核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.38
langgraph>=0.0.26

# 向量数据库
chromadb>=0.4.22

# 工具与测试
requests>=2.31.0
pytest>=7.4.0
python-dotenv>=1.0.0
